Rename @types to types
