+++ src/types/artist.ts
