+++ src/types/image_type.ts
